refactor: 修复越权/商品上架链路并收敛重复代码
- 修复收货地址删除越权(去 BaseMapper.deleteById, 按归属删)
- 商品上架回填主键并返回 id, 打通 上架->初始化库存 链路; Feign 结果显式校验
- 订单业务失败改抛 BusinessException 返回友好提示
- 新增 mall-common: IdentityInterceptor/Feign 公共配置/Rabbit 拓扑常量, 删 3 份死 JwtUtil 及各服务重复拦截器
- 主键/外键全链路 Integer->Long(与 BIGINT 对齐)
- 删除无效布隆过滤器代码; 去除伪乐观锁 version 自增; 移除 bootstrap/seata/guava 依赖
- 下单扣库存 MQ 增加 Redis SETNX 幂等; 清死日志与注释残留
- 重写 README: 补主要业务逻辑

+++ mall-gateway/src/main/java/com/gateway/config/AuthGlobalFilter.java
@@ -50,7 +50,7 @@ public class AuthGlobalFilter implements GlobalFilter, Ordered {
             if (idObj == null) {
                 throw new RuntimeException("token 缺少用户 id");
             }
-            Integer id = ((Number) idObj).intValue();
+            long id = ((Number) idObj).longValue();
             String username = (String) claims.get("username");
 
             // 校验登录态仍有效（单设备登录 / 主动失效）
