+++ mall-service/mall-service-product/src/main/java/com/product/mapper/CategoryMapper.java
@@ -10,7 +10,7 @@ public interface CategoryMapper {
 
     //浏览：某父分类下的启用子分类，按 sort_order 升序
     @Select("select * from category where parent_id=#{parentId} and status=1 order by sort_order,id")
-    List<Category> listEnabledByParent(@Param("parentId") Integer parentId);
+    List<Category> listEnabledByParent(@Param("parentId") Long parentId);
 
     //浏览：全部启用分类（供拼树），按父/排序
     @Select("select * from category where status=1 order by parent_id,sort_order,id")
@@ -21,13 +21,13 @@ public interface CategoryMapper {
     List<Category> listAll();
 
     @Select("select * from category where id=#{id}")
-    Category findById(@Param("id") Integer id);
+    Category findById(@Param("id") Long id);
 
     //同一父分类下是否已存在同名分类（新增时 excludeId 传 0；编辑时排除自身）
     @Select("select count(*) from category where parent_id=#{parentId} and name=#{name} and id<>#{excludeId}")
-    long countSiblingByName(@Param("parentId") Integer parentId,
+    long countSiblingByName(@Param("parentId") Long parentId,
                             @Param("name") String name,
-                            @Param("excludeId") Integer excludeId);
+                            @Param("excludeId") Long excludeId);
 
     @Insert("insert into category(parent_id,name,sort_order,status,created_time,updated_time) " +
             "values(#{parentId},#{name},#{sortOrder},#{status},now(),now())")
