+++ mall-service/mall-service-product/src/main/java/com/product/service/ProductNumService.java
@@ -7,11 +7,11 @@ public interface ProductNumService {
 
     void addNumProduct(Product product);
 
-    Product findNumProductByUserIdAndName(Integer userId, String name);
+    Product findNumProductByUserIdAndName(Long userId, String name);
 
     /** 商家编辑自己商品（部分更新，仅改传入字段），带归属校验 */
-    void updateProduct(Integer userId, UpdateProductRequest request);
+    void updateProduct(Long userId, UpdateProductRequest request);
 
     /** 商家上/下架自己商品：status=1 上架 / 0 下架 */
-    void changeProductStatus(Integer userId, Integer id, Integer status);
+    void changeProductStatus(Long userId, Long id, Integer status);
 }
