+++ mall-service/mall-service-product/src/main/java/com/product/service/impl/CategoryServiceImpl.java
@@ -22,8 +22,8 @@ public class CategoryServiceImpl implements CategoryService {
     CategoryMapper categoryMapper;
 
     @Override
-    public List<Category> listEnabledByParent(Integer parentId) {
-        int p = (parentId == null || parentId < 0) ? 0 : parentId;
+    public List<Category> listEnabledByParent(Long parentId) {
+        Long p = (parentId == null || parentId < 0) ? 0L : parentId;
         return categoryMapper.listEnabledByParent(p);
     }
 
@@ -33,16 +33,16 @@ public class CategoryServiceImpl implements CategoryService {
         if (all == null || all.isEmpty()) {
             return List.of();
         }
-        Map<Integer, List<Category>> byParent = new HashMap<>();
+        Map<Long, List<Category>> byParent = new HashMap<>();
         for (Category c : all) {
-            int pid = (c.getParentId() == null) ? 0 : c.getParentId();
+            long pid = (c.getParentId() == null) ? 0L : c.getParentId();
             byParent.computeIfAbsent(pid, k -> new ArrayList<>()).add(c);
         }
         byParent.values().forEach(kids -> kids.sort(Comparator.comparing(Category::getSortOrder,
                 Comparator.nullsFirst(Integer::compareTo)).thenComparing(Category::getId)));
 
-        List<Category> roots = byParent.getOrDefault(0, List.of());
-        Set<Integer> visited = new HashSet<>();
+        List<Category> roots = byParent.getOrDefault(0L, List.of());
+        Set<Long> visited = new HashSet<>();
         roots.forEach(root -> attach(root, byParent, visited));
         return roots;
     }
@@ -51,7 +51,7 @@ public class CategoryServiceImpl implements CategoryService {
     public Category add(Category category) {
         normalize(category);
         validateParent(category.getParentId(), null);
-        if (categoryMapper.countSiblingByName(category.getParentId(), category.getName(), 0) > 0) {
+        if (categoryMapper.countSiblingByName(category.getParentId(), category.getName(), 0L) > 0) {
             throw new BusinessException("同级已存在同名分类");
         }
         categoryMapper.insertCategory(category);
@@ -79,7 +79,7 @@ public class CategoryServiceImpl implements CategoryService {
         categoryMapper.updateCategory(category);
     }
 
-    private void attach(Category node, Map<Integer, List<Category>> byParent, Set<Integer> visited) {
+    private void attach(Category node, Map<Long, List<Category>> byParent, Set<Long> visited) {
         if (!visited.add(node.getId())) {
             node.setChildren(null); // 防御历史脏数据造成环
             return;
@@ -94,14 +94,14 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     private void normalize(Category c) {
-        if (c.getParentId() == null) c.setParentId(0);
+        if (c.getParentId() == null) c.setParentId(0L);
         if (c.getSortOrder() == null) c.setSortOrder(0);
         if (c.getStatus() == null) c.setStatus(1);
     }
 
     /** 父分类必须为 0 或已存在；编辑时禁止把分类挂到自己之下 */
-    private void validateParent(Integer parentId, Integer selfId) {
-        if (parentId == 0) {
+    private void validateParent(Long parentId, Long selfId) {
+        if (parentId == 0L) {
             return;
         }
         if (parentId != null && parentId.equals(selfId)) {
